Distributed multi container fix
