+++ .gitignore
@@ -1,1 +1,1 @@
-gatling/results/
+gatling/2.3.0/results/
