+++ gatling/2.3.0/gatling.conf
@@ -1,0 +1,10 @@
+gatling {
+  data {
+    writers = [file, graphite]      # The list of DataWriters to which Gatling write simulation data (currently supported : console, file, graphite, jdbc)
+    graphite {
+      host = "influxdb"         # The host where the Carbon server is located
+      port = 2003                # The port to which the Carbon server listens to (2003 is default for plaintext, 2004 is default for pickle)
+      protocol = "tcp"           # The protocol used to send data to Carbon (currently supported : "tcp", "udp")
+    }
+  }
+}
