+++ influxdb/1.2.0-alpine/Dockerfile
@@ -1,0 +1,5 @@
+FROM influxdb:1.2.0-alpine
+MAINTAINER ugurarpaci
+ADD influxdb.conf /etc/influxdb/influxdb.conf
+ENTRYPOINT ["/entrypoint.sh"]
+CMD ["influxd"]
