--- influxdb/influxdb.conf
@@ -1,86 +1,0 @@
-reporting-disabled = false
-bind-address = ":8088"
-hostname = ""
-join = ""
-
-[meta]
-  dir = "/var/lib/influxdb/meta"
-  retention-autocreate = true
-  logging-enabled = true
-  pprof-enabled = false
-  lease-duration = "1m0s"
-
-[data]
-  dir = "/var/lib/influxdb/data"
-  engine = "tsm1"
-  wal-dir = "/var/lib/influxdb/wal"
-  wal-logging-enabled = true
-  query-log-enabled = true
-  cache-max-memory-size = 524288000
-  cache-snapshot-memory-size = 26214400
-  cache-snapshot-write-cold-duration = "1h0m0s"
-  compact-full-write-cold-duration = "24h0m0s"
-  max-points-per-block = 0
-  data-logging-enabled = true
-
-[retention]
-  enabled = true
-  check-interval = "30m0s"
-
-[shard-precreation]
-  enabled = true
-  check-interval = "10m0s"
-  advance-period = "30m0s"
-
-[admin]
-  enabled = true
-  bind-address = ":8083"
-  https-enabled = false
-  https-certificate = "/etc/ssl/influxdb.pem"
-  Version = ""
-
-[monitor]
-  store-enabled = true
-  store-database = "_internal"
-  store-interval = "10s"
-
-[subscriber]
-  enabled = true
-
-[http]
-  enabled = true
-  bind-address = ":8086"
-  auth-enabled = false
-  log-enabled = true
-  write-tracing = false
-  pprof-enabled = false
-  https-enabled = false
-  https-certificate = "/etc/ssl/influxdb.pem"
-  max-row-limit = 10000
-
-[[graphite]]
-  enabled = true
-  bind-address = ":2003"
-  database = "graphite"
-  protocol = "tcp"
-  batch-size = 5000
-  batch-pending = 10
-  batch-timeout = "1s"
-  consistency-level = "one"
-  separator = "."
-  udp-read-buffer = 0
-# https://github.com/influxdata/influxdb/blob/master/services/graphite/README.md  
-#  templates = [
-#    "gatling.*.*.*.count measurement.simulation.request.status.field",
-#    "gatling.*.*.*.min measurement.simulation.request.status.field",
-#    "gatling.*.*.*.max measurement.simulation.request.status.field",
-#    "gatling.*.*.*.percentiles50 measurement.simulation.request.status.field",
-#    "gatling.*.*.*.percentiles75 measurement.simulation.request.status.field",
-#    "gatling.*.*.*.percentiles95 measurement.simulation.request.status.field",
-#    "gatling.*.*.*.percentiles99 measurement.simulation.request.status.field"
-#  ]
-
-[continuous_queries]
-  log-enabled = true
-  enabled = true
-  run-interval = "1s"
